Rename 'published' attribute to 'created' in blog posts
