+++ website/posts/challenges-levels-rewards-and-punishments.en.md
@@ -2,7 +2,7 @@
 title: Challenges, Levels, Rewards and Punishments
 subtitle: Nature may offer beyond your current awareness if you care to feel its wholeness.
 image: /images/blog/sunset.jpg
-published: 2025-11-14
+created: 2025-11-14
 updated: 2026-02-26
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org
