+++ website/posts/definition-of-love.en.md
@@ -1,8 +1,9 @@
 ---
 title: Definition of Love
 subtitle: Love is a unique adventure in each mind but everybody crosses similar paths.
-published: 2026-
+created: 2026-
 image: /images/blog/-.jpg
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org 
+published: false
 ---
