+++ website/posts/definition-of-love.tr.md
@@ -1,7 +1,7 @@
 ---
 title: Aşkın Tanımı
 subtitle: Herkesin zihninde yaşadığı aşk serüveni farklıdır ama benzer yollarda kesişirler.
-published: 2026-
+created: 2026-
 image: /images/blog/-.jpg
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org
