+++ website/posts/modern-kingdoms.en.md
@@ -1,7 +1,7 @@
 ---
 title: Modern Kingdoms
 subtitle: There are different management styles and regimes in the world, but they are all structured as modern kingdoms in different sizes.
-published: 2026-
+created: 2026-
 image: /images/blog/-.jpg
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org
