+++ website/posts/modern-kingdoms.tr.md
@@ -1,7 +1,7 @@
 ---
 title: Çağımızın Krallıkları
 subtitle: Dünyanın her yerinde farklı yönetim şekilleri ve rejimleri bulunuyor, fakat hepsi farklı boyutlardaki modern krallık yapılardır.
-published: 2026-
+created: 2026-
 image: /images/blog/-.jpg
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org
