+++ website/posts/reading-habit.en.md
@@ -1,7 +1,7 @@
 ---
 title: Reading Habit
 subtitle: The act of reading is usually binded to alphabets, words, or books. We need a broader focus on its meaning because we don't know how to read nature.
-published: 2026-
+created: 2026-
 image: /images/blog/-.jpg
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org
