+++ website/posts/reading-habit.tr.md
@@ -1,7 +1,7 @@
 ---
 title: Okuma Alışkanlığı
 subtitle: Okuma eylemi genellikle alfabelere, kelimelere ve kitaplara yönelik kullanılıyor. Bu bağlamda daha geniş bir bakış açısına ihtiyacımız var, çünkü doğayı okumayı bilmiyoruz.
-published: 2026-
+created: 2026-
 image: /images/blog/-.jpg
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org
