+++ website/posts/stories-and-myths.en.md
@@ -1,7 +1,7 @@
 ---
 title: Stories and Myths
 subtitle: Every culture has different stories and myths that reveal the concepts of their minds. By connecting their details, we can be aware of similarities.
-published: 2026-
+created: 2026-
 image: /images/blog/-.jpg
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org
