+++ website/posts/stories-and-myths.tr.md
@@ -1,7 +1,7 @@
 ---
 title: Hikayeler ve Mitolojiler
 subtitle: Her kültürün kendi zihin yapısını yansıttığı farklı hikayeler ve mitolojiler bulunuyor. Detaylarının bağlanmasıyla birlikte benzerlikleri fark edebiliriz.
-published: 2026-
+created: 2026-
 image: /images/blog/-.jpg
 authorName: Emin Buğra Saral
 authorUrl: https://0.1234S.org
