+++ website/src/app/[locale]/blog/[slug]/page.tsx
@@ -1,5 +1,5 @@
 import BlogPost from '@/app/components/Pages/BlogPost';
-import { getPost } from '@/lib/posts';
+import { getPost } from '@/app/lib/posts';
 import { getIntlayer, getMultilingualUrls } from 'intlayer';
 import type { Metadata } from 'next';
 import { LocalPromiseParams, type NextPageIntlayer } from 'next-intlayer';
